We now generate "write" serialization code for structs
It's not perfect; indentation is busted, and the way we define/track struct members needs a lot of cleanup. However, we do get functional enough results now.

---

Needed changes in future commits:

`member` should probably be an abstract class, with `integral_member`, `string_member`, `struct_member`, and `inlined_union_member` as subtypes.

`c_type` should be folded into those; array rank stuff should be moved to the `member` base class.

Alignments should be moved to `member`; bitfield stuff should be moved to `integral_member`; the "struct" versus "union" declaration (for unions that aren't inline-declared) should be moved to `struct_member`.

`c_type::primitive` should be a freestanding enum and should be renamed to `integral_type`.

`member::explicit_bitcount` should be renamed to `serialization_bitcount` to distinguish it from a member that is a bitfield in RAM, and it should be moved to `integral_member`.

When doing C codegen, we should use a state object with member functions for managing indentation.

The `ast::structure` class should have member functions for listing all struct dependencies, checking if any directly-contained member is a string, etc..

We need the code for generating lu_BitstreamRead_ClassNameHere functions.

For `string_member` instances, the max length of the string should not be appended to the array extent list. Rather, we should just check for it when generating struct-definition output.