We now track preprocessor constants mirrored in the XML
When a field refers to a preprocessor constant in some way (e.g. array length, bitfield size, etc.), we now remember which constant it referred to and emit the constant name in our generated C code. Additionally, generated code uses preprocessor directives to double-check that constant values have not changed since the code was last generated.