Whole-struct read code; coalescing of paused arrays in sectors
We now generate code to bitstream-read whole structs.

Previously, if an array with six elements were split at a sector boundary after its third element (index 2), we'd serialize array[0], array[1], and array[2] individually before the split, and array[3], array[4], and array[5] after the split. Now, we generate for-loops for each half of the array. We do this within the sector generator: upon encountering an array item, we check the items following it to see if they're "array siblings" (indices defined out to the same rank; last defined index has a difference of +1), and if so, we swallow all those siblings and emit a loop for them.

XML definitions: global item IDs are now 9-bit, allowing us to make saved inventory data quite a bit smaller (half the size of vanilla).