Rebuilt the program with a GUI
Created a `codegen-gui` project, as part of work to make this whole thing easier to share to other Emerald hack projects.

Sector groups are no longer hardcoded; they can be defined in XML; see `_manifest.xml`.

We now generate save functors and the tables for them, i.e. the body of `sSaveSlotLayout` in my savegame code and the declarations and definitions of functions like `ReadSector_WorldData`.

The program now has a GUI.

Moved the bitstream code to the `lu` namespace and cleaned it up a bit. Documented more of it, too.

Savegame dumper functionality is now integrated.