fixes
The most notable is that we now validate all struct and union names even inside of non-serialized union branches. We have to: since we no longer use `c-type-decl`, we need a struct/union definition in order to know whether the type is defined via typedef, which affects how we refer to it in generated C code.